# Conflicts:
#	target/classes/KursovProektOOP2/data/entity/Formular.class
#	target/classes/KursovProektOOP2/data/entity/Warehouse.class
#	target/classes/KursovProektOOP2/data/repository/OwnerRepository.class
#	target/classes/KursovProektOOP2/data/repository/WarehouseRepository.class

+++ src/main/java/KursovProektOOP2/data/repository/ProductTypeRepository.java
@@ -64,7 +64,7 @@ public class ProductTypeRepository implements DAORepository{
     }
 
     @Override
-    public Optional getById(Long id) {
+    public Optional getById(int id) {
         Session session = Connection.openSession();
         Transaction transaction = session.beginTransaction();
         Optional<ProductType> foundProductType = null;
